requirements-dev.txt
-r ./requirements.txt
pytest_homeassistant_custom_component~=0.13.39
pip>=21.0,<23.2
ruff==0.0.275
colorlog==6.7.0
ruamel.yaml==0.17.32
aiohttp_cors==0.7.0